+++ src/coreclr/nativeaot/Runtime/unix/PalCreateDump.h
@@ -8,4 +8,4 @@
 extern bool PalCreateDumpInitialize();
 extern void PalCreateCrashDumpIfEnabled();
 extern void PalCreateCrashDumpIfEnabled(int signal, siginfo_t* siginfo);
-extern void PalCreateCrashDumpIfEnabled(void* pExAddress, void* pExContext, void* triageBuffer, int triageBufferSize);
+extern void PalCreateCrashDumpIfEnabled(void* pExceptionRecord, void* pExContext);
